🛢 Create endpoint for submitting project for review
Currently it uses the same serializer as the Task create-PR logic, which
makes it easier to factor but does differ from the XD mockup. I'd like
to discuss whether there are reasons to have them use different input
serializers.
